+++ src/main/java/com/rimo/sfcr/mixin/extra/ServerLevelMixin.java
@@ -21,7 +21,7 @@ public abstract class ServerLevelMixin {
 			target = "Lnet/minecraft/world/level/biome/Biome;shouldSnow(Lnet/minecraft/world/level/LevelReader;Lnet/minecraft/core/BlockPos;)Z"
 	))
 	private boolean sfcr$stopSnowLayerPileUp(Biome instance, LevelReader level, BlockPos pos, Operation<Boolean> original) {
-		if (Common.isNoCloudCovered((Level) level, pos.getX(), pos.getY(), pos.getZ()))
+		if (Common.CONFIG.isCloudRainLogically() && Common.isNoCloudCovered((Level) level, pos.getX(), pos.getY(), pos.getZ()))
 			return false;
 		return original.call(instance, level, pos);
 	}
@@ -32,7 +32,7 @@ public abstract class ServerLevelMixin {
 			target = "Lnet/minecraft/world/level/biome/Biome;getPrecipitation()Lnet/minecraft/world/level/biome/Biome$Precipitation;"
 	))
 	private Biome.Precipitation sfcr$stopCauldronPileUp(Biome instance, Operation<Biome.Precipitation> original, @Local(ordinal = 1) BlockPos pos) {
-		if (Common.isNoCloudCovered((Level) (Object) this, pos.getX(), pos.getY(), pos.getZ()))
+		if (Common.CONFIG.isCloudRainLogically() && Common.isNoCloudCovered((Level) (Object) this, pos.getX(), pos.getY(), pos.getZ()))
 			return Biome.Precipitation.NONE;
 		return original.call(instance);
 	}
@@ -40,7 +40,7 @@ public abstract class ServerLevelMixin {
 	//~ if < 1.21.1 'tickPrecipitation' -> 'tickChunk'
 	@ModifyVariable(method = "tickPrecipitation", at = @At("STORE"))
 	private Biome.Precipitation sfcr$stopCauldronPileUp(Biome.Precipitation precipitation, @Local(ordinal = 1) BlockPos pos) {
-		if (Common.isNoCloudCovered((Level) (Object) this, pos.getX(), pos.getY(), pos.getZ()))
+		if (Common.CONFIG.isCloudRainLogically() && Common.isNoCloudCovered((Level) (Object) this, pos.getX(), pos.getY(), pos.getZ()))
 			return Biome.Precipitation.NONE;
 		return precipitation;
 	}